Cria funções de ETL para os dados que irão compor o painel de produtividade do AGP
